refactor(observability): promote //src/observability Bazel package (DEV-6496)
Step (d) of Probe-9a — first per-module Bazel package promotion.
`src/observability/` becomes its own `cc_library` with co-located unit test
per ADR-0003.

## What changed

`src/observability/BUILD.bazel` (new):
  cc_library "observability"
    srcs: connection_metrics_adapter.cpp, metrics.cpp, sentry.cpp, sentry_init.cpp
    hdrs: connection_metrics_adapter.h, metrics.h, sentry.h, sentry_init.h
    deps: //ext/sentry, //include:headers, //shttps:shttps,
          //src:sipi_generated_hdrs, @prometheus_cpp//core
    visibility: //src:__subpackages__ + //shttps:__pkg__ (the latter
      documents the known Metrics::instance() consumption from
      shttps/Server.cpp per CONTEXT.md)
  cc_test "connection_metrics_adapter_test"
    srcs: connection_metrics_adapter_test.cpp + test_main.cpp (moved
      from test/unit/sipiconnectionmetrics/)
    deps: :observability + @googletest//:gtest_main

`src/BUILD.bazel`:
- New cc_library `:sipi_generated_hdrs` wraps `generated/{SipiVersion,SipiConfig,*_icc}.h`
  so per-module cc_libraries can depend on them without pulling
  `:sipi_lib`'s archive (which would close a cycle — `:sipi_lib` itself
  depends on `:observability`).
- `:sipi_lib` loses its inline `hdrs` list of generated files; gains
  `:sipi_generated_hdrs` and `//src/observability:observability` in deps.
- Drop `**/*.h` from the srcs glob added in PR 1 — subpackage boundary
  now excludes `src/observability/*.h` from the parent glob anyway, and
  no other top-level src/ files use `.h`.

`src/populate_from_image.cpp` (new, under :sipi_lib):
- Out-of-line implementation of `Sipi::observability::populate_from_image`,
  which dereferences `Sipi::SipiImage`. Co-locating it in
  `src/observability/sentry.cpp` would force `:observability` to depend
  on whatever target owns SipiImage (today `:sipi_lib`), closing a
  circular dependency. Move back into `observability/sentry.cpp` once
  SipiImage gets its own package per DEV-6388 / DEV-6395.
- File-static helpers `predefined_profile_to_string` /
  `orientation_to_string` move with it (only this TU calls them).

`src/observability/sentry.cpp`:
- Drops `populate_from_image` + the two helpers + the `SipiImage.hpp` /
  `metadata/SipiIcc.h` includes.
- Now depends only on `<sys/stat.h>` and the public `observability/sentry.h`.

`src/observability/connection_metrics_adapter_test.cpp` + `test_main.cpp`:
- Moved from `test/unit/sipiconnectionmetrics/` per ADR-0003. The legacy
  `test/unit/sipiconnectionmetrics/BUILD.bazel` is deleted.

`justfile`:
- `bazel-test`, `bazel-test-unit`, `bazel-coverage` recipes now also
  test `//src/...` so per-module ADR-0003 co-located tests get picked
  up automatically as more modules promote.

## Layering check deferred

`features = ["layering_check", "parse_headers"]` is **not** enabled here
yet, despite the DEV-6496 spec calling for it. Same disposition as
`//shttps:shttps`'s deferred enablement (Logger.h rehome): waiting on
SipiImage decomposition (DEV-6388) so the includes below resolve cleanly:
- `sentry.h` forward-declares `Sipi::SipiImage`; the strict layering
  check would still allow the fwd-decl, but the symmetry with the rest
  of the codebase is cleaner once SipiImage has a real home.
- DEV-6353 is the global flip-the-switch PR that will turn these features
  on for every module that has settled by then.

## Acceptance

- bazel build //src:sipi (and //src/observability:observability): green
- bazel test //src/observability:connection_metrics_adapter_test: PASSED
- bazel test //test/unit/...: 11 tests pass (the adapter test now lives
  under //src/observability — picked up by the updated bazel-test-unit
  recipe)
- bazel test //test/approval:approvaltests: PASSED

Refs DEV-6489.
