+++ src/observability/connection_metrics_adapter_test.cpp
