--- test/unit/sipiconnectionmetrics/BUILD.bazel
@@ -1,14 +1,0 @@
-"""SipiConnectionMetricsAdapter unit tests — verifies the adapter
-forwards shttps connection-lifecycle events to Sipi::observability::Metrics.
-"""
-
-load("@rules_cc//cc:defs.bzl", "cc_test")
-
-cc_test(
-    name = "sipi_connection_metrics_adapter_test",
-    srcs = glob(["*.cpp"]),
-    deps = [
-        "//src:sipi_lib",
-        "@googletest//:gtest_main",
-    ],
-)
